requirements.txt
flask
pandas
mysql-connector-python
openpyxl
xlrd==1.2.0
matplotlib